requirements.txt
python==3.8.12
pytorch==1.10.0
pyqt5==5.15.4
six==1.16.0
tqdm==4.62.3
onnx==1.10.2
numpy==1.22.1
imgaug==0.4.0
einops==0.3.2
pandas==1.3.3
pip
visdom==0.1.8.9
imageio==2.9.0

sklearn
IPython==7.29.0
matplotlib==3.4.3

tensorboardX==2.2
json
opencv-python==4.5.3.56
openpyxl==3.0.9